+++ ui/src/components/menu/menu-icon.ts
@@ -22,7 +22,7 @@ import { ICONS } from '../core/shared/icons';
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
   exportAs: 'ouiMenuIcon',
-  // tslint:disable-next-line:no-host-metadata-property
+  // eslint-disable-next-line @angular-eslint/no-host-metadata-property
   host: {
     '(menuOpened)': 'menuOpened()',
     '(menuClosed)': 'menuClosed()',