+++ ui/src/components/paginator/paginator.ts
@@ -37,6 +37,7 @@ export class PageEvent {
 
   /**
    * Index of the page that was selected previously.
+   *
    * @breaking-change 8.0.0 To be made into a required property.
    */
   previousPageIndex?: number;
@@ -65,9 +66,9 @@ export const _OuiPaginatorBase: CanDisableCtor &
   exportAs: 'ouiPaginator',
   templateUrl: 'paginator.html',
   styleUrls: ['paginator.scss'],
-  // tslint:disable-next-line:no-inputs-metadata-property
+  // eslint-disable-next-line @angular-eslint/no-inputs-metadata-property
   inputs: ['disabled'],
-  // tslint:disable-next-line:no-host-metadata-property
+  // eslint-disable-next-line @angular-eslint/no-host-metadata-property
   host: {
     class: 'oui-paginator',
   },
@@ -123,9 +124,8 @@ export class OuiPaginator
   private _hidePageSize = false;
 
   /** Event emitted when the paginator changes the page size or page index. */
-  @Output() readonly page: EventEmitter<PageEvent> = new EventEmitter<
-    PageEvent
-  >();
+  @Output()
+  readonly page: EventEmitter<PageEvent> = new EventEmitter<PageEvent>();
 
   /** Displayed set of page size options. Will be sorted and include current page size. */
   _displayedPageSizeOptions: number;