+++ ui/src/components/tooltip/tooltip.ts
@@ -50,6 +50,7 @@ export const TOOLTIP_PANEL_CLASS = 'oui-tooltip-panel';
 
 /**
  * Creates an error to be thrown if the user supplied an invalid tooltip position.
+ *
  * @docs-private
  */
 export function getOuiTooltipInvalidPositionError(position: string) {
@@ -83,13 +84,18 @@ export interface OuiTooltipDefaultOptions {
   touchendHideDelay: number;
 }
 
+export interface NewCSSStyleDeclaration extends CSSStyleDeclaration {
+  msUserSelect: string;
+}
+
 /** Injection token to be used to override the default options for `ouiTooltip`. */
-export const OUI_TOOLTIP_DEFAULT_OPTIONS = new InjectionToken<
-  OuiTooltipDefaultOptions
->('oui-tooltip-default-options', {
-  providedIn: 'root',
-  factory: OUI_TOOLTIP_DEFAULT_OPTIONS_FACTORY,
-});
+export const OUI_TOOLTIP_DEFAULT_OPTIONS = new InjectionToken<OuiTooltipDefaultOptions>(
+  'oui-tooltip-default-options',
+  {
+    providedIn: 'root',
+    factory: OUI_TOOLTIP_DEFAULT_OPTIONS_FACTORY,
+  }
+);
 
 /** @docs-private */
 export function OUI_TOOLTIP_DEFAULT_OPTIONS_FACTORY(): OuiTooltipDefaultOptions {
@@ -104,6 +110,7 @@ export type TooltipVisibility = 'initial' | 'visible' | 'hidden';
 
 /**
  * Internal component that wraps the tooltip's content.
+ *
  * @docs-private
  */
 @Component({
@@ -113,7 +120,7 @@ export type TooltipVisibility = 'initial' | 'visible' | 'hidden';
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
   animations: [ouiTooltipAnimations.tooltipState],
-  // tslint:disable-next-line:no-host-metadata-property
+  // eslint-disable-next-line @angular-eslint/no-host-metadata-property
   host: {
     // Forces the element to have a layout in IE and Edge. This fixes issues where the element
     // won't be rendered if the animations are disabled or there is no web animations polyfill.
@@ -155,6 +162,7 @@ export class TooltipComponent {
 
   /**
    * Shows the tooltip with an animation originating from the provided origin
+   *
    * @param delay Amount of milliseconds to the delay showing the tooltip.
    */
   show(): void {
@@ -178,6 +186,7 @@ export class TooltipComponent {
 
   /**
    * Begins the animation to hide the tooltip after the provided delay in ms.
+   *
    * @param delay Amount of milliseconds to delay showing the tooltip.
    */
   hide(): void {
@@ -247,7 +256,7 @@ export class TooltipComponent {
 @Directive({
   selector: '[ouiTooltip]',
   exportAs: 'ouiTooltip',
-  // tslint:disable-next-line:no-host-metadata-property
+  // eslint-disable-next-line @angular-eslint/no-host-metadata-property
   host: {
     '(longpress)': 'show()',
     '(keydown)': '_handleKeydown($event)',
@@ -368,7 +377,7 @@ export class OuiTooltip implements OnDestroy, CanDisable {
   ) {
     this._scrollStrategy = scrollStrategy;
     const element: HTMLElement = _elementRef.nativeElement;
-    const elementStyle = element.style as CSSStyleDeclaration & {
+    const elementStyle = element.style as NewCSSStyleDeclaration & {
       webkitUserDrag: string;
     };
     const hasGestures = typeof window === 'undefined' || (window as any).Hammer;